refactor : 구간등록 로직  도메인객체로 이동
